Updates 0.5.1
Added lores subroutines and macros, as well as most of the documentation for it. You will not find documentation for the library yet in the quick reference area.

At this point, it is likely that I will be going back to revisit older disks in order to do more optimization, make some strategic file name changes, and add extra subroutines/macros to each library. More content will be added to the documentation as well in order to accomodate multiple audiences.

After that, we will be moving to version 0.6.0, at which point the speaker library and the hires library will be added.